Рефакторинг моделей и добавление статистических методов сервисов

+++ src/admin/logs/models.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from admin.logs.enums import ELogType
 from database import Base
 from sqlalchemy.orm import mapped_column, Mapped
-from sqlalchemy import JSON, Enum, Integer, String, DateTime, func, ForeignKey
+from sqlalchemy import JSON, Enum, Integer, String, DateTime, func, Float
 
 class Log(Base):
     __tablename__ = "log"
@@ -11,7 +11,7 @@ class Log(Base):
     created_at: Mapped[DateTime] = mapped_column(DateTime, server_default=func.now())
     offer_id: Mapped[str] = mapped_column(String)
     skins: Mapped[dict] = mapped_column(JSON)
-    total_price: Mapped[int] = mapped_column(Integer, nullable=True)
+    total_price: Mapped[float] = mapped_column(Float, nullable=True)
     status: Mapped[ELogType] = mapped_column(Enum(ELogType))
     target_steam_id: Mapped[str] = mapped_column(String)
     bot_steam_id: Mapped[str] = mapped_column(String)
